fix(fuzz): widen fuzz_lottery_validation stored_pool to u128
validate_block_lottery and compute_pool_accounting now take the
cumulative carryover as u128, so the fuzz target's FuzzData.stored_pool
must match (it was u64). Widening it also lets the fuzzer exercise pool
values above u64::MAX. Keeps the workspace-excluded fuzz crate building
under the pinned cargo-fuzz toolchain (Fuzz Build Guard).

Closes #343

+++ fuzz/fuzz_targets/fuzz_lottery_validation.rs
@@ -52,7 +52,10 @@ struct FuzzData {
     /// Raw bytes decoded into a Block (arbitrary lottery_summary / outputs).
     block_bytes: Vec<u8>,
     candidates: Vec<FuzzCandidate>,
-    stored_pool: u64,
+    // Cumulative carryover pool is u128 (widened from u64 to prevent
+    // saturation under sustained inflow); fuzz the full range, including
+    // values above u64::MAX.
+    stored_pool: u128,
     prev_block_hash: [u8; 32],
 }
